Adjusting Bass service to support newly created interfaces with doubles at the input.

+++ src/SoundFingerprinting.Audio.Bass/BassAudioService.cs
@@ -40,7 +40,7 @@
             }
         }
 
-        public override AudioSamples ReadMonoSamplesFromFile(string pathToSourceFile, int sampleRate, int seconds, int startAt)
+        public override AudioSamples ReadMonoSamplesFromFile(string pathToSourceFile, int sampleRate, double seconds, double startAt)
         {
             int stream = streamFactory.CreateStream(pathToSourceFile);
             float[] samples = resampler.Resample(stream, sampleRate, seconds, startAt, mixerStream => new BassSamplesProvider(proxy, mixerStream));
